requirements.txt
streamlit
openai
pandas
faiss-cpu
numpy
sentence-transformers
scikit-learn
